Added Signal Chests, working on ore destroyers
Finished

Got done with another new type of chest, a signal chest! These are an addon to the direction chests.
If a signal chest does not have the new item, a signal, the direction chest wont move objects.
Signal chests have to be next to a direction chest to work.

The reason why I created these is to add on to more ways to automate direction chests and, hopefully, progress more onto making them turing complete.
Need to research more into what needs to happen to make them turing complete though :D

Working on

I'm currently working on Ore Destroyers, objects that destroy rocks that are near in a 5 by 5 area and store the resource that they provide. Supposed to be used near a quarry.
Currently on an issue due to a stack overflow- how did I mess up that bad

For the future

1. Wood Destroyers. They are like Ore Destroyers but for trees.

2. Crop Harvesters. They are like Ore Destroyers but for your crops! Auto crop harvester, if you will.